+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [2.5.3] - 04-08-2026
+
+### Added
+- Import examples how to filter data for account and memberships
+
+### Fixed
+- Added private key check in scripts where it was missing
+
 ## [2.5.2] - 30-07-2026
 ### Fixed
 - **correlateOnly/create.ps1**: Changed correlation value from `accountFieldValue` to `personFieldValue` to support Governance correlation scenarios correctly.